remove webview message handling from experiments
